+++ src/SaltyEngine/SFML/EventManager.cpp
@@ -84,12 +84,12 @@ namespace SaltyEngine {
 
         ::SaltyEngine::Vector EventManager::GetPosition(void) {
             sf::Vector2i position = sf::Mouse::getPosition();
-            return ::SaltyEngine::Vector(position.x, position.y);
+            return ::SaltyEngine::Vector(static_cast<float>(position.x), static_cast<float>(position.y));
         }
 
         ::SaltyEngine::Vector EventManager::GetPositionRelative(void) {
             sf::Vector2i position = sf::Mouse::getPosition(*m_window);
-            return ::SaltyEngine::Vector(position.x, position.y);
+            return ::SaltyEngine::Vector(static_cast<float>(position.x), static_cast<float>(position.y));
         }
 
         bool EventManager::IsJoystickConnected(unsigned int id) {
